fix: 회원가입 에러 수정

+++ feature/signup/src/main/java/com/combo/runcombi/signup/model/PetInfoData.kt
@@ -1,6 +1,6 @@
 package com.combo.runcombi.signup.model
 
 data class PetInfoData(
-    val petWeight: Int? = null,
+    val petWeight: Double? = null,
     val petAge: Int? = null,
 )
